Indicate caps lock state on the login widget
Not everyone has a keyboard with an obvious caps lock indicator, and
even if the message box tells you to check the caps lock, not everyone
is smart enough to actually do that. So, create an indicator (just a
label) that watches the caps lock. It can query the state, since it may
not have been active when the key was pressed, but you can drive it
yourself if you want.