Change APP_STYLESHEET_END to APP_STYLESHEET_STOP to match the theme files.
The code has been looking for APP_STYLESHEET_END, whereas the themes
have been using APP_STYLESHEET_STOP. It works the way that it's been,
because the code also checks for the end of the file, and
APP_STYLESHEET_STOP is at the end of the file, but they really should
match. I'm not sure that it really matters whether it's END or STOP, but
I went with STOP, because that meant changing one line of code instead
of changing all of the themes.